+++ src/lib.rs
@@ -13,6 +13,7 @@ pub mod asap;
 pub mod auth;
 pub mod bt;
 pub mod coefficient;
+pub mod content_class;
 pub mod curator;
 pub mod db;
 pub mod db_health;